added margin for report borders

+++ src/Common/Eclipse.Common/Plots/AxisOptions.cs
@@ -4,6 +4,6 @@ public sealed class AxisOptions<TValue>
 {
     public TValue[] Values { get; set; } = [];
     public string Label { get; set; } = string.Empty;
-    public int Min { get; set; }
-    public int Max { get; set; }
+    public double Min { get; set; }
+    public double Max { get; set; }
 }
